Add device creation via REST. Not tested

+++ iot-services/src/main/java/com/jmssolutions/iot/dao/AbstractJpaDAO.java
@@ -3,9 +3,7 @@ package com.jmssolutions.iot.dao;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.persistence.TypedQuery;
+import javax.persistence.*;
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
@@ -63,7 +61,9 @@ public abstract class AbstractJpaDAO<K, E> implements DataAccessObject<K, E> {
     @Transactional
     public void remove(K id) {
         E e = (E)entityManager.find(entityClass, id);
-        entityManager.remove(e);
+        if(e != null)
+            entityManager.remove(e);
+        else throw new EntityNotFoundException();
     }
 
 }
